extracts stencil matrix generation into separate file
